Fix ServerShared for CultMath
CultMath's hash(float2) returns float, not uint, so ZoneGenerator folds its
bits into the zone seed with BitConverter.SingleToInt32Bits. MathResolver's
int2?[] entry used ArrayFormatter<float2?>; it now uses ArrayFormatter<int2?>.

Aetheria.Shared and AetherDb build headless with 0 errors, and AetherDb's
census, factions and hardpoint-fit output is byte-identical to the captures
taken before the swap.

+++ Assets/Scripts/ServerShared/CultCache/Serialization/MathResolver.cs
@@ -52,7 +52,7 @@ internal static class MathResolverResolverGetFormatterHelper
         { typeof(float3[]), new ArrayFormatter<float3>() },
         { typeof(float4[]), new ArrayFormatter<float4>() },
         { typeof(float2?[]), new ArrayFormatter<float2?>() },
-        { typeof(int2?[]), new ArrayFormatter<float2?>() },
+        { typeof(int2?[]), new ArrayFormatter<int2?>() },
         { typeof(float3?[]), new ArrayFormatter<float3?>() },
         { typeof(float4?[]), new ArrayFormatter<float4?>() },
